FFmpeg video decoder glue code Implementation of FFmpeg to perform video decoding. Also added hooks to insert this FFmpeg video decoder filter.
Review URL: http://codereview.chromium.org/60069

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@13056 0039d316-1c4b-4281-b951-d872f2087c98

+++ media/base/buffers.h
@@ -157,8 +157,9 @@ struct VideoSurface {
   size_t planes;
 
   // Array of strides for each plane, typically greater or equal to the width
-  // of the surface divided by the horizontal sampling period.
-  size_t strides[kMaxPlanes];
+  // of the surface divided by the horizontal sampling period.  Note that
+  // strides can be negative.
+  int32 strides[kMaxPlanes];
 
   // Array of data pointers to each plane.
   uint8* data[kMaxPlanes];
